@@ -1,8 +1,13 @@
 package com.sid.gl.model;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
 import java.util.Collection;
 
+@Getter
+@Setter
 @Entity
 @Table(name="tb_privilege")
 public class Privilege {
